requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0

langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
langsmith>=0.0.77
pydantic>=2.0.0

numpy>=1.24.0
scikit-learn>=1.3.0
networkx>=3.0
matplotlib>=3.7.0
requests>=2.31.0

beautifulsoup4>=4.12.0
duckduckgo-search>=6.0.0
fake-useragent>=1.5.0
selenium>=4.20.0

tqdm>=4.66.0
colorama>=0.4.6
